refactor(project): FSD 레이어 위반 수정
- ProjectActionMenu를 features/delete-project에서 views/mypage로 이동
  - 두 feature(edit, delete)를 조합하는 컴포넌트는 상위 레이어에 귀속
  - features/delete-project/index.ts에 DeleteConfirmModal export 추가
- register-project 슬라이스를 project-form으로 리네임
  - 등록/수정 모드를 모두 담당하는 슬라이스 책임 명확화
  - RegisterProjectForm → ProjectForm으로 컴포넌트명 변경
