+++ src/features/project-form/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectForm } from './ui/ProjectForm';
