+++ src/features/project-form/lib/getCroppedImg.ts
