+++ src/features/project-form/ui/ParticipantsField.tsx
