--- src/features/register-project/index.ts
@@ -1,1 +1,0 @@
-export { default as RegisterProjectForm } from './ui/RegisterProjectForm';
